refactored ui tooling
