+++ src/shared/hooks/index.ts
@@ -2,5 +2,6 @@ export * from "./useCustom";
 export * from "./useFormContext";
 export * from "./useFormCtx";
 export * from "./useForms";
+export * from "./useIcons";
 export * from "./useToast";
 export * from "./useTooltip";
