--- src/tw-styled/stories/Styles.demo.tsx
@@ -1,13 +1,0 @@
-"use client";
-
-export const StylesDemo = () => {
-  return (
-    <div className="demo">
-      <div>
-        <div className="demo w-xs" />
-        <div className="demo w-xs" />
-        <div className="demo w-xs" />
-      </div>
-    </div>
-  );
-};
