--- src/tw-styled/theme/useThemeCtx.ts
@@ -1,8 +1,0 @@
-import React from "react";
-import { ThemeContext } from "./ThemeProvider";
-
-export const useThemeCtx = () => {
-  const theme = React.useContext(ThemeContext);
-
-  return theme;
-};
