+++ src/tw-styled/tools.ts
@@ -1,3 +1,1 @@
-export * from "./utils/class-merge";
-export * from "./hooks/useIcons";
 export * from "./style-to-class-resolver";
