--- src/tw-styled/types/Children.ts
@@ -1,1 +1,0 @@
-export type Children = React.ReactNode;
