+++ src/tw-styled/types/index.ts
@@ -1,5 +1,3 @@
-export * from "./Children";
 export * from "./Elements";
 export * from "./Theming";
-
 export * from "../style-to-class-resolver/types";
