+++ src/utils/theme/ThemeProvider.tsx
@@ -1,5 +1,4 @@
 import React from "react";
-import { Children } from "../types";
 import { themeCtx, ThemeCtx } from "./themeCtx";
 
 export const ThemeContext = React.createContext<ThemeCtx>(themeCtx);
@@ -9,7 +8,13 @@ export const ThemeProvider = ({
   children
 }: {
   theme: any;
-  children: Children;
+  children: React.ReactNode;
 }) => {
   return <ThemeContext.Provider value={theme}>{children}</ThemeContext.Provider>;
+};
+
+export const useThemeCtx = () => {
+  const theme = React.useContext(ThemeContext);
+
+  return theme;
 };
