+++ src/utils/theme/index.ts
@@ -1,3 +1,2 @@
 export * from "./ThemeProvider";
 export * from "./themeCtx";
-export * from "./useThemeCtx";
